requirements.txt
PyMuPDF==1.23.14
pytesseract
pdf2image
pillow
scikit-learn
